requirements.txt
Flask
flask_cors
numpy
tensorflow
torch
torchvision
opencv-python-headless
Pillow
werkzeug
matplotlib
onnx
tqdm
gunicorn